+++ src/ui/ScreenList.cpp
@@ -22,7 +22,7 @@ vui::IGameScreen* vui::ScreenList::movePrevious() {
 void vui::ScreenList::setScreen(i32 s) {
     _current = s;
 }
-vui::ScreenList* vui::ScreenList::addScreens(IGameScreen** s, i32 c /*= 0*/) {
+vui::ScreenList* vui::ScreenList::addScreens(IGameScreen** s, size_t c /*= 0*/) {
     // Check For Arguments
     if (s == 0 || c < 1) return this;
 
@@ -45,16 +45,16 @@ vui::ScreenList* vui::ScreenList::addScreens(IGameScreen** s, i32 c /*= 0*/) {
 vui::ScreenList* vui::ScreenList::addScreen(IGameScreen* s) {
     _screens.push_back(s);
     if (s != 0) {
-        s->setParentGame(_game, _screens.size() - 1);
+        s->setParentGame(_game, (i32)_screens.size() - 1);
         s->build();
     }
     return this;
 }
 
 void vui::ScreenList::destroy(GameTime gameTime) {
-    int s = _screens.size();
+    i32 s = (i32)_screens.size();
     if (s < 1) return;
-    for (int i = 0; i < s; i++) {
+    for (i32 i = 0; i < s; i++) {
         _screens[i]->destroy(gameTime);
     }
     _screens.resize(0);
